Rename numeric_types.hpp to numeric.hpp
