+++ include/BURST/numeric.hpp
