+++ include/BURST/robot.hpp
@@ -12,7 +12,7 @@
 #include <boost/container/small_vector.hpp>
 
 #include "geometry.hpp"
-#include "numeric_types.hpp"
+#include "numeric.hpp"
 #include "renderable.hpp"
 #include "configuration_space.hpp"
 #include "models.hpp"
